# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jams": {
      "post": {
        "operationId": "JamController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Jam created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JamResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new jam session",
        "tags": [
          "Jams"
        ]
      },
      "get": {
        "operationId": "JamController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of jams",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/JamResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "skip": {
                          "type": "number"
                        },
                        "take": {
                          "type": "number"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List all jams with pagination",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}": {
      "get": {
        "operationId": "JamController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jam found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JamResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "summary": "Get jam by ID",
        "tags": [
          "Jams"
        ]
      },
      "patch": {
        "operationId": "JamController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jam updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update jam",
        "tags": [
          "Jams"
        ]
      },
      "delete": {
        "operationId": "JamController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jam deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete jam",
        "tags": [
          "Jams"
        ]
      }
    },
...
    "/jams/{id}/control/start": {
      "post": {
        "operationId": "JamController_startJam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jam started successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start jam playback (starts first scheduled song)",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/control/stop": {
      "post": {
        "operationId": "JamController_stopJam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jam stopped successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stop jam playback",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/control/next": {
      "post": {
        "operationId": "JamController_nextSong",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skipped to next song"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Skip to next song",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/control/previous": {
      "post": {
        "operationId": "JamController_previousSong",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned to previous song"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Go back to previous song",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/control/pause": {
      "post": {
        "operationId": "JamController_pauseSong",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song paused"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pause current song",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/control/resume": {
      "post": {
        "operationId": "JamController_resumeSong",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song resumed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resume paused song",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/control/reorder": {
      "post": {
        "operationId": "JamController_reorderSchedules",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderSchedulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queue reordered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JamResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid schedule IDs or validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not jam host"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reorder schedules in jam queue (drag and drop support)",
        "tags": [
          "Jams"
        ]
      }
    },
    "/jams/{id}/playback-history": {
      "get": {
        "operationId": "JamController_getPlaybackHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playback history retrieved"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get playback history for jam",
        "tags": [
          "Jams"
        ]
      }
    },
...
  },
  "info": {
    "title": "Karaoke Jam API",
    "description": "Backend API para organização de Jam Sessions presenciais em tempo real",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateJamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Jam session name"
          },
          "description": {
            "type": "string",
            "description": "Jam session description"
          },
          "date": {
            "type": "string",
            "description": "Jam session date and time"
          },
          "location": {
            "type": "string",
            "description": "Jam session location"
          },
          "hostMusicianId": {
            "type": "string",
            "description": "Host musician ID (UUID)"
          },
          "hostName": {
            "type": "string",
            "description": "Host name (denormalized cache)"
          },
          "hostContact": {
            "type": "string",
            "description": "Host contact (denormalized cache)"
          },
          "qrCode": {
            "type": "string",
            "description": "Jam session QR Code"
          },
          "status": {
            "type": "string",
            "description": "Jam session status",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LIVE",
              "FINISHED"
            ]
          }
        },
        "required": [
          "name",
          "location"
        ]
      },
      "MusicaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "neededDrums": {
            "type": "number",
            "example": 0,
            "description": "Number of drummers needed"
          },
          "neededGuitars": {
            "type": "number",
            "example": 0,
            "description": "Number of guitarists needed"
          },
          "neededVocals": {
            "type": "number",
            "example": 0,
            "description": "Number of vocalists needed"
          },
          "neededBass": {
            "type": "number",
            "example": 0,
            "description": "Number of bassists needed"
          },
          "neededKeys": {
            "type": "number",
            "example": 0,
            "description": "Number of keyboardists needed"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "status",
          "neededDrums",
          "neededGuitars",
          "neededVocals",
          "neededBass",
          "neededKeys",
          "createdAt"
        ]
      },
      "MusicianResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "instrument": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "supabaseUserId": {
            "type": "string"
          },
          "isHost": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isHost",
          "createdAt"
        ]
      },
      "RegistrationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "musicianId": {
            "type": "string"
          },
          "jamId": {
            "type": "string"
          },
          "jamMusicId": {
            "type": "string"
          },
          "scheduleId": {
            "type": "string"
          },
          "instrument": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "musician": {
            "$ref": "#/components/schemas/MusicianResponseDto"
          }
        },
        "required": [
          "id",
          "musicianId",
          "jamId",
          "jamMusicId",
          "status",
          "createdAt"
        ]
      },
      "JamMusicaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jamId": {
            "type": "string"
          },
          "musicId": {
            "type": "string"
          },
          "music": {
            "$ref": "#/components/schemas/MusicaResponseDto"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponseDto"
            }
          }
        },
        "required": [
          "id",
          "jamId",
          "musicId",
          "music"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jamId": {
            "type": "string"
          },
          "musicId": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "SUGGESTED",
              "SCHEDULED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "pausedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "music": {
            "$ref": "#/components/schemas/MusicaResponseDto"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponseDto"
            }
          }
        },
        "required": [
          "id",
          "jamId",
          "musicId",
          "order",
          "status",
          "createdAt"
        ]
      },
      "JamResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "hostName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "qrCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LIVE",
              "FINISHED"
            ]
          },
          "playbackState": {
            "type": "string",
            "enum": [
              "STOPPED",
              "PLAYING",
              "PAUSED"
            ]
          },
          "currentScheduleId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "jamMusics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JamMusicaResponseDto"
            }
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponseDto"
            }
          },
          "schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ScheduleOrderUpdateDto": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "string",
            "description": "Schedule ID",
            "format": "uuid"
          },
          "order": {
            "type": "number",
            "description": "New order position (1-based)",
            "minimum": 1
          }
        },
        "required": [
          "scheduleId",
          "order"
        ]
      },
      "ReorderSchedulesDto": {
        "type": "object",
        "properties": {
          "updates": {
            "description": "Array of schedule order updates",
            "example": [
              {
                "scheduleId": "123e4567-e89b-12d3-a456-426614174000",
                "order": 1
              },
              {
                "scheduleId": "123e4567-e89b-12d3-a456-426614174001",
                "order": 2
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleOrderUpdateDto"
            }
          }
        },
        "required": [
          "updates"
        ]
      },
      "UpdateJamDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Jam status",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LIVE",
              "FINISHED"
            ]
          }
        }
      },
...
    }
  }
}